Refs #127. Split up the styles module
