+++ pydocx/openxml/wordprocessing/styles.py
@@ -1,3 +1,4 @@
+# coding: utf-8
 from __future__ import (
     absolute_import,
     print_function,
@@ -6,48 +7,7 @@ from __future__ import (
 
 from collections import defaultdict
 
-from pydocx.models import XmlModel, ChildTag, Attribute
-from pydocx.types import OnOff, Underline
-
-
-class RunProperties(XmlModel):
-    bold = ChildTag(type=OnOff, name='b', attrname='val')
-    italic = ChildTag(type=OnOff, name='i', attrname='val')
-    underline = ChildTag(type=Underline, name='u', attrname='val')
-    caps = ChildTag(type=OnOff, attrname='val')
-    small_caps = ChildTag(type=OnOff, name='smallCaps', attrname='val')
-    strike = ChildTag(type=OnOff, attrname='val')
-    dstrike = ChildTag(type=OnOff, attrname='val')
-    vanish = ChildTag(type=OnOff, attrname='val')
-    hidden = ChildTag(type=OnOff, name='webHidden', attrname='val')
-    vertical_align = ChildTag(name='vertAlign', attrname='val')
-    parent_style = ChildTag(name='rStyle', attrname='val')
-    pos = ChildTag(name='position', attrname='val')
-    sz = ChildTag(name='sz', attrname='val')
-
-    @property
-    def position(self):
-        if self.pos is None:
-            return 0
-        return int(self.pos)
-
-    @property
-    def size(self):
-        if self.sz is None:
-            return
-        return int(self.sz)
-
-
-class ParagraphProperties(XmlModel):
-    parent_style = ChildTag(name='pStyle', attrname='val')
-
-
-class Style(XmlModel):
-    style_type = Attribute(name='type', default='paragraph')
-    style_id = Attribute(name='styleId', default='')
-    name = ChildTag(attrname='val', default='')
-    run_properties = ChildTag(type=RunProperties, name='rPr')
-    parent_style = ChildTag(name='basedOn', attrname='val')
+from pydocx.openxml.wordprocessing.style import Style
 
 
 class Styles(object):
